reorganize dependencies and packages
