+++ src/main/kotlin/com/chsdngm/tilly/repository/VoteRepository.kt
@@ -1,10 +1,10 @@
-package com.imbananko.tilly.repository
+package com.chsdngm.tilly.repository
 
-import com.imbananko.tilly.model.MemeEntity
-import com.imbananko.tilly.model.MemeStatsEntry
-import com.imbananko.tilly.model.VoteEntity
-import com.imbananko.tilly.model.VoteValue.*
-import com.imbananko.tilly.utility.SqlQueries
+import com.chsdngm.tilly.model.MemeEntity
+import com.chsdngm.tilly.model.MemeStatsEntry
+import com.chsdngm.tilly.model.VoteEntity
+import com.chsdngm.tilly.model.VoteValue.*
+import com.chsdngm.tilly.utility.SqlQueries
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate
 import org.springframework.stereotype.Repository
