+++ src/main/kotlin/com/chsdngm/tilly/utility/BotConfig.kt
@@ -1,4 +1,4 @@
-package com.imbananko.tilly.utility
+package com.chsdngm.tilly.utility
 
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Component
